+++ db/app/cfg/src/full/R__14.PACKAGE_BODY.CFG_INSTALL_PKG.sql
@@ -151,6 +151,23 @@ begin
       l_plsql_flags := l_plsql_flags || ',Testing:true';
   end;
   
+  -- It must be possible to install PATO in a database without APEX installed.
+  
+  begin
+    execute immediate 'begin :v := WWV_FLOW_API.C_CURRENT; end;' using out l_found;
+
+    if l_found >= 20100513
+    then
+      l_plsql_flags := l_plsql_flags || ',APEX:true';
+    else
+      raise no_data_found;
+    end if;
+  exception
+    when others
+    then
+      l_plsql_flags := l_plsql_flags || ',APEX:false';
+  end;
+
   if l_plsql_flags is not null
   then
     l_plsql_flags := ltrim(l_plsql_flags, ',');